Dimension order (#21)
* Update tokenizer for standardized dimension order.

Also added tests. They will need to be updated when we add multi-segment tokenization (or we'll add new tests for that).

* Standardize dimensions.

inputs x tokens x embeddings

* Update documentation.

* Update copyright.

* Update vignette.

* Bump dev vnum.

So I feel like we're acknowledging that this is a big change.

* Simplify tests.

Down with transposing! Since the previous tests passed, I accepted the outputs here as true and updated the expectations to match.

* Add note to README about the dev version of torch.